add package named xml and move files in package
