+++ Spring1/src/main/java/org/example/xml/Alien.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.xml;
 
 public class Alien {
     private int age;
