+++ Spring1/src/main/java/org/example/xml/App.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.xml;
 
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
