+++ Spring1/src/main/java/org/example/xml/Computer.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.xml;
 
 public interface Computer {
     public void compile();
