+++ Spring1/src/main/java/org/example/xml/Desktop.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.xml;
 
 public class Desktop implements Computer {
     @Override
