+++ Spring1/src/main/java/org/example/xml/Laptop.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.xml;
 
 public class Laptop implements Computer {
     public Laptop() {
