+++ Spring1/src/main/resources/spring.xml
@@ -5,12 +5,12 @@
 
 <!--  By default, the scope is 'singleton' on bean tag  -->
 
-    <bean id="alien" class="org.example.Alien" autowire="byType">
+    <bean id="alien" class="org.example.xml.Alien" autowire="byType">
 <!--   In the attribute 'name', the instance variable name i.e. age should be provided. Then, it will look for the 'setter method' i.e. setAge() of 'age' variable, and pass the value: 27 through parameter of setter method. This is how setter injection works.     -->
         <property name="age" value="27" />
         <property name="salary" value="100000000" />
         <property name="computer">
-            <bean id="computer" class="org.example.Laptop" primary="true" />
+            <bean id="computer" class="org.example.xml.Laptop" primary="true" />
         </property>
 
 <!--    This constructor tag is used if value is to be passed through param constructor in the 'Alien' bean (class).     -->
@@ -19,10 +19,10 @@
 <!--        <constructor-arg index="2" ref="laptop" />-->
     </bean>
 
-<!--    <bean id="computer" class="org.example.Laptop" primary="true" >-->
+<!--    <bean id="computer" class="org.example.xml.Laptop" primary="true" >-->
 <!--    </bean>-->
 
-<!--    <bean id="computer2" class="org.example.Desktop">-->
+<!--    <bean id="computer2" class="org.example.xml.Desktop">-->
 <!--    </bean>-->
 
 </beans>
